requirements.txt
streamlit==1.2.0
tensorflow-cpu==2.6.0
Pillow==8.3.2
pandas==1.3.5